requirements.txt
pyodbc
sqlalchemy
boto3
pandas
pyyaml
plotly
dash
pyarrow
s3fs
geopandas
shapely
geopy
joblib
psutil